do not move if duplicate file detected

+++ main.py
@@ -6,6 +6,9 @@ def loopFolder(folderPath):
     for subdir, dirs, files in os.walk(folderPath):
         for file in files:
             # print(os.path.join(subdir, file))
-            shutil.move(os.path.join(subdir, file), folderPath)
+            if not os.path.exists(os.path.join(folderPath, file)):
+                shutil.move(os.path.join(subdir, file), folderPath)
+            else:
+                print("The file destination_file already exists.")            
     print('done!')
 loopFolder("D:\\toTest")
